consultar roles y permisos. Ocultar o mostrar botones y enlaces, dependiendo de los permisos del usuario autenticado

+++ resources/views/dosis/tabla.php
@@ -1,3 +1,8 @@
+<?php
+
+    use App\Auth;
+
+?>
 <div class="table-responsive">
     <table id="tabla" class="table table-striped" tyle="width: 100%;">
         <thead>
@@ -17,6 +22,7 @@
                     <td><?php echo $value->vacuna->nombre  ?></td>
                     <td><?php echo $value->fecha_aplicacion  ?></td>
                     <td>
+                    <?php if(Auth::hasPermission('dosis.eliminar')): ?>
                         <a 
                             onclick="return confirm('¿Dese eliminar a esta dosis?')" 
                             href="dosis/delete/<?php echo $value->id  ?>" 
@@ -24,6 +30,7 @@
                         >
                             <i class="fas fa-trash"></i>
                         </a>
+                    <?php endif ?>
                     </td>
                 </tr>
             <?php endforeach ?>
